feat: use dot product and limit allowed files to 16b wave files

+++ dtmf/src/fft.c
@@ -2,6 +2,7 @@
 #include "utils.h"
 #include <stdbool.h>
 #include <stddef.h>
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -30,7 +31,7 @@ int fft(cplx_t *buf, size_t n)
 	return 0;
 }
 
-void float_to_cplx_t(const int32_t *in, cplx_t *out, size_t n)
+void float_to_cplx_t(const int16_t *in, cplx_t *out, size_t n)
 {
 	for (size_t i = 0; i < n; ++i) {
 		out[i] = in[i] + 0. * I;
